+++ lib/src/parse/visitor.dart
@@ -57,6 +57,7 @@ class SpecWalker {
     for (final response in operation.responses.responses.values) {
       _refOr(response);
     }
+    _maybeRefOr(operation.responses.defaultResponse);
     for (final parameter in operation.parameters) {
       _refOr(parameter);
     }